requirements.txt
aiodns==3.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
audioop-lts==0.2.1
brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
customtkinter==5.2.2
darkdetect==0.8.0
discord-py==2.5.2
dotenv==0.9.9
frozenlist==1.7.0
grpcio==1.73.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
msgspec==0.19.0
multidict==6.4.4
packaging==25.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
py-cord==2.6.1
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycares==4.9.0
pycparser==2.22
pydantic==2.11.5
pydantic-core==2.33.2
pynacl==1.5.0
pyparsing==3.2.3
python-dotenv==1.1.0
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
uritemplate==4.2.0
urllib3==2.4.0
wavelink==3.4.1
websockets==15.0.1
yarl==1.20.1
